+++ lizard_portal/media/lizard_portal/lizard/popup/TimeSeriesGraph.js
@@ -11,11 +11,10 @@
     init_background: null,
     statics: {
       show: function(records, workspaceitem) {
-        var bbar, collage_item_config, collage_item_identifier, dt_end, dt_start, record, title;
+        var bbar, collage_item_config, collage_item_identifier, csv_html, dt_end, dt_start, graph_item_html, graph_title, img_html, qua_ident_extra, record, single_record, title, _i, _len;
         dt_start = Ext.Date.format(Lizard.CM.getContext().period.start, 'Y-m-d H:i:s');
         dt_end = Ext.Date.format(Lizard.CM.getContext().period.end, 'Y-m-d H:i:s');
         record = records[0];
-        title = workspaceitem.get('text') + ' - ' + record.data.geo_ident;
         collage_item_identifier = {
           geo_ident: record.data.geo_ident,
           par_ident: record.data.par_ident,
@@ -25,15 +24,19 @@
           fews_norm_source_slug: record.data.fews_norm_source_slug
         };
         collage_item_config = {
-          name: workspaceitem.get('text') + ' - ' + record.data.geo_ident,
-          title: workspaceitem.get('text') + ' - ' + record.data.geo_ident,
+          name: workspaceitem.get('title') + ' - ' + record.data.geo_ident,
+          title: workspaceitem.get('title') + ' - ' + record.data.geo_ident,
           plid: workspaceitem.get('plid'),
           js_popup_class: workspaceitem.get('js_popup_class'),
-          identifier: Ext.JSON.encode(collage_item_identifier)
+          identifier: Ext.JSON.encode(collage_item_identifier),
+          grouping_hint: 'tijdreeks ' + record.data.par_ident
         };
+        graph_title = 'Grafiek voor ' + record.data.geo_ident + ' ' + record.data.par_ident + ' ' + record.data.mod_ident + ' ' + record.data.stp_ident;
         if (record.data.is_collage_item === true) {
+          title = 'Collage popup voor ' + record.data.grouping_hint;
           bbar = [];
         } else {
+          title = workspaceitem.get('title') + ' - ' + record.data.geo_ident;
           bbar = [
             {
               text: 'Voeg toe aan collage',
@@ -45,6 +48,18 @@
             }
           ];
         }
+        graph_item_html = '';
+        for (_i = 0, _len = records.length; _i < _len; _i++) {
+          single_record = records[_i];
+          if (single_record.data.qua_ident) {
+            qua_ident_extra = '%22qua_ident%22:%22' + single_record.data.qua_ident + '%22';
+          } else {
+            qua_ident_extra = '';
+          }
+          graph_item_html += '&item={%22fews_norm_source_slug%22:%22' + single_record.data.fews_norm_popup_slug + '%22,%22location%22:%22' + single_record.data.geo_ident + '%22,%22parameter%22:%22' + single_record.data.par_ident + '%22,%22type%22:%22line%22,%22time_step%22:%22' + single_record.data.stp_ident + '%22,%22module%22:%22' + single_record.data.mod_ident + '%22' + qua_ident_extra + '}';
+        }
+        img_html = '<img src="/graph/?dt_start=' + dt_start + '&dt_end=' + dt_end + '&width=1000&height=500' + graph_item_html + '" />';
+        csv_html = '<a href="/graph/?dt_start=' + dt_start + '&dt_end=' + dt_end + graph_item_html + '&format=csv" >csv downloaden</a>';
         return Ext.create('Ext.window.Window', {
           title: title,
           modal: true,
@@ -57,7 +72,7 @@
               xtype: 'panel',
               width: 1050,
               height: 550,
-              html: 'Grafiek voor ' + record.data.geo_ident + ' ' + record.data.par_ident + ' ' + record.data.mod_ident + ' ' + record.data.stp_ident + '<img src="/graph/?dt_start=' + dt_start + '&dt_end=' + dt_end + '&width=1000&height=500&item={%22fews_norm_source_slug%22:%22waternet%22,%22location%22:%22' + record.data.geo_ident + '%22,%22parameter%22:%22' + record.data.par_ident + '%22,%22type%22:%22line%22,%22time_step%22:%22' + record.data.stp_ident + '%22,%22module%22:%22' + record.data.mod_ident + '%22}" />',
+              html: '<h3>' + graph_title + '</h3><br />' + img_html + '<br /> ' + csv_html,
               bbar: bbar
             }
           ]